removed device_utils package
